Rename DefaultProperty, make it just for table property defaults
