+++ java/core/src/main/java/sleeper/core/properties/instance/CommonProperty.java
@@ -183,6 +183,23 @@ public interface CommonProperty {
             .validationPredicate(value -> SleeperPropertyValueUtils.isListWithMaxSize(value, 5))
             .runCdkDeployWhenChanged(true)
             .propertyGroup(InstancePropertyGroup.COMMON).build();
+    UserDefinedInstanceProperty DEFAULT_LAMBDA_CONCURRENCY_RESERVED = Index.propertyBuilder("sleeper.default.lambda.concurrency.reserved")
+            .description("Default value for the reserved concurrency for each lambda within the Sleeper instance. " +
+                    "By default no concurrency is reserved for the lambdas. Each lambda also has its own property " +
+                    "that overrides the value found here.\n" +
+                    "See reserved concurrency overview at: https://docs.aws.amazon.com/lambda/latest/dg/configuration-concurrency.html")
+            .validationPredicate(SleeperPropertyValueUtils::isPositiveIntegerOrNull)
+            .defaultValue(null)
+            .propertyGroup(InstancePropertyGroup.COMMON).build();
+    UserDefinedInstanceProperty DEFAULT_LAMBDA_CONCURRENCY_MAXIMUM = Index.propertyBuilder("sleeper.default.lambda.concurrency.max")
+            .description("Default value for the maximum concurrency for each lambda within the Sleeper instance. " +
+                    "By default the maximum concurrency is set to 10, which is enough for 10 online tables. " +
+                    "If there are more online tables, this number may need to be increased. Each lambda also has its own property that " +
+                    "overrides the value found here.\n" +
+                    "See maximum concurrency overview at: https://aws.amazon.com/blogs/compute/introducing-maximum-concurrency-of-aws-lambda-functions-when-using-amazon-sqs-as-an-event-source/")
+            .validationPredicate(SleeperPropertyValueUtils::isPositiveIntegerOrNull)
+            .defaultValue("10")
+            .propertyGroup(InstancePropertyGroup.COMMON).build();
 
     static List<UserDefinedInstanceProperty> getAll() {
         return Index.INSTANCE.getAll();
