+++ java/core/src/main/java/sleeper/core/properties/instance/TableDefaultProperty.java
@@ -30,71 +30,71 @@ import java.util.Locale;
 import static sleeper.core.properties.validation.SleeperPropertyValueUtils.describeEnumValuesInLowerCase;
 
 /**
- * Definitions of instance properties that are defaults for other properties, particularly table properties.
+ * Definitions of instance properties that are defaults for table properties.
  */
-public interface DefaultProperty {
+public interface TableDefaultProperty {
     UserDefinedInstanceProperty DEFAULT_S3A_READAHEAD_RANGE = Index.propertyBuilder("sleeper.default.fs.s3a.readahead.range")
             .description("The readahead range set on the Hadoop configuration when reading Parquet files in a query\n" +
                     "(see https://hadoop.apache.org/docs/current/hadoop-aws/tools/hadoop-aws/index.html).")
             .defaultValue("64K")
             .validationPredicate(SleeperPropertyValueUtils::isValidHadoopLongBytes)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_ROW_GROUP_SIZE = Index.propertyBuilder("sleeper.default.rowgroup.size")
             .description("The size of the row group in the Parquet files (default is 8MiB).")
             .defaultValue("" + (8 * 1024 * 1024)) // 8 MiB
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_PAGE_SIZE = Index.propertyBuilder("sleeper.default.page.size")
             .description("The size of the pages in the Parquet files (default is 128KiB).")
             .defaultValue("" + (128 * 1024)) // 128 KiB
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_COMPRESSION_CODEC = Index.propertyBuilder("sleeper.default.compression.codec")
             .description("The compression codec to use in the Parquet files.\n" +
                     "Valid values are: " + describeEnumValuesInLowerCase(CompressionCodec.class))
             .defaultValue("zstd")
             .validationPredicate(CompressionCodec::isValid)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_DICTIONARY_ENCODING_FOR_ROW_KEY_FIELDS = Index.propertyBuilder("sleeper.default.parquet.dictionary.encoding.rowkey.fields")
             .description("Whether dictionary encoding should be used for row key columns in the Parquet files.")
             .defaultValue("false")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_DICTIONARY_ENCODING_FOR_SORT_KEY_FIELDS = Index.propertyBuilder("sleeper.default.parquet.dictionary.encoding.sortkey.fields")
             .description("Whether dictionary encoding should be used for sort key columns in the Parquet files.")
             .defaultValue("false")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_DICTIONARY_ENCODING_FOR_VALUE_FIELDS = Index.propertyBuilder("sleeper.default.parquet.dictionary.encoding.value.fields")
             .description("Whether dictionary encoding should be used for value columns in the Parquet files.")
             .defaultValue("false")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_COLUMN_INDEX_TRUNCATE_LENGTH = Index.propertyBuilder("sleeper.default.parquet.columnindex.truncate.length")
             .description("Used to set parquet.columnindex.truncate.length, see documentation here:\n" +
                     "https://github.com/apache/parquet-mr/blob/master/parquet-hadoop/README.md\n" +
                     "The length in bytes to truncate binary values in a column index.")
             .defaultValue("128")
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_STATISTICS_TRUNCATE_LENGTH = Index.propertyBuilder("sleeper.default.parquet.statistics.truncate.length")
             .description("Used to set parquet.statistics.truncate.length, see documentation here:\n" +
                     "https://github.com/apache/parquet-mr/blob/master/parquet-hadoop/README.md\n" +
                     "The length in bytes to truncate the min/max binary values in row groups.")
             .defaultValue("2147483647")
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_PARQUET_WRITER_VERSION = Index.propertyBuilder("sleeper.default.parquet.writer.version")
             .description("Used to set parquet.writer.version, see documentation here:\n" +
                     "https://github.com/apache/parquet-mr/blob/master/parquet-hadoop/README.md\n" +
                     "Can be either v1 or v2. The v2 pages store levels uncompressed while v1 pages compress levels with the data.")
             .defaultValue("v2")
             .validationPredicate(List.of("v1", "v2")::contains)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_ADD_TRANSACTION_MAX_ATTEMPTS = Index.propertyBuilder("sleeper.default.statestore.transactionlog.add.transaction.max.attempts")
             .description("The number of attempts to make when applying a transaction to the state store. " +
                     "This default can be overridden by a table property.")
             .defaultValue("" + TransactionLogStateStore.DEFAULT_MAX_ADD_TRANSACTION_ATTEMPTS)
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_ADD_TRANSACTION_FIRST_RETRY_WAIT_CEILING_MS = Index.propertyBuilder("sleeper.default.statestore.transactionlog.add.transaction.first.retry.wait.ceiling.ms")
             .description("The maximum amount of time to wait before the first retry when applying a transaction to " +
                     "the state store. Full jitter will be applied so that the actual wait time will be a random " +
@@ -104,7 +104,7 @@ public interface DefaultProperty {
                     "This default can be overridden by a table property.")
             .defaultValue("" + TransactionLogStateStore.DEFAULT_RETRY_WAIT_RANGE.getFirstWaitCeiling().toMillis())
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_ADD_TRANSACTION_MAX_RETRY_WAIT_CEILING_MS = Index.propertyBuilder("sleeper.default.statestore.transactionlog.add.transaction.max.retry.wait.ceiling.ms")
             .description("The maximum amount of time to wait before any retry when applying a transaction to " +
                     "the state store. Full jitter will be applied so that the actual wait time will be a random " +
@@ -114,7 +114,7 @@ public interface DefaultProperty {
                     "This default can be overridden by a table property.")
             .defaultValue("" + TransactionLogStateStore.DEFAULT_RETRY_WAIT_RANGE.getMaxWaitCeiling().toMillis())
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_TIME_BETWEEN_SNAPSHOT_CHECKS_SECS = Index.propertyBuilder("sleeper.default.statestore.transactionlog.time.between.snapshot.checks.seconds")
             .description("The number of seconds to wait after we've loaded a snapshot before looking for a new " +
                     "snapshot. This should relate to the rate at which new snapshots are created, configured in the " +
@@ -122,7 +122,7 @@ public interface DefaultProperty {
                     "This default can be overridden by a table property.")
             .defaultValue("" + TransactionLogStateStore.DEFAULT_TIME_BETWEEN_SNAPSHOT_CHECKS.toSeconds())
             .validationPredicate(SleeperPropertyValueUtils::isNonNegativeInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_TIME_BETWEEN_TRANSACTION_CHECKS_MS = Index.propertyBuilder("sleeper.default.statestore.transactionlog.time.between.transaction.checks.ms")
             .description("The number of milliseconds to wait after we've updated from the transaction log before " +
                     "checking for new transactions. The state visible to an instance of the state store can be out " +
@@ -132,18 +132,18 @@ public interface DefaultProperty {
                     "This default can be overridden by a table property.")
             .defaultValue("" + TransactionLogStateStore.DEFAULT_TIME_BETWEEN_TRANSACTION_CHECKS.toMillis())
             .validationPredicate(SleeperPropertyValueUtils::isNonNegativeInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_MIN_TRANSACTIONS_AHEAD_TO_LOAD_SNAPSHOT = Index.propertyBuilder("sleeper.default.statestore.transactionlog.snapshot.load.min.transactions.ahead")
             .description("The minimum number of transactions that a snapshot must be ahead of the local " +
                     "state, before we load the snapshot instead of updating from the transaction log.")
             .defaultValue("" + TransactionLogStateStore.DEFAULT_MIN_TRANSACTIONS_AHEAD_TO_LOAD_SNAPSHOT)
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_TRANSACTION_LOG_SNAPSHOT_EXPIRY_IN_DAYS = Index.propertyBuilder("sleeper.default.statestore.transactionlog.snapshot.expiry.days")
             .description("The number of days that transaction log snapshots remain in the snapshot store before being deleted.")
             .defaultValue("2")
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT)
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT)
             .build();
     UserDefinedInstanceProperty DEFAULT_TRANSACTION_LOG_SNAPSHOT_MIN_AGE_MINUTES_TO_DELETE_TRANSACTIONS = Index
             .propertyBuilder("sleeper.default.statestore.transactionlog.delete.behind.snapshot.min.age.minutes")
@@ -154,7 +154,7 @@ public interface DefaultProperty {
                     "reading the log.")
             .defaultValue("2")
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT)
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT)
             .build();
     UserDefinedInstanceProperty DEFAULT_TRANSACTION_LOG_NUMBER_BEHIND_TO_DELETE = Index.propertyBuilder("sleeper.default.statestore.transactionlog.delete.number.behind.latest.snapshot")
             .description("The minimum number of transactions that a transaction must be behind the latest snapshot " +
@@ -171,20 +171,20 @@ public interface DefaultProperty {
                     "sleeper.default.statestore.transactionlog.delete.behind.snapshot.min.age\n")
             .defaultValue("200")
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT)
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT)
             .build();
     UserDefinedInstanceProperty DEFAULT_DYNAMO_STRONGLY_CONSISTENT_READS = Index.propertyBuilder("sleeper.default.table.dynamo.strongly.consistent.reads")
             .description("This specifies whether queries and scans against DynamoDB tables used in the state stores " +
                     "are strongly consistent. This default can be overridden by a table property.")
             .defaultValue("false")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_BULK_IMPORT_MIN_LEAF_PARTITION_COUNT = Index.propertyBuilder("sleeper.default.bulk.import.min.leaf.partitions")
             .description("Specifies the minimum number of leaf partitions that are needed to run a bulk import job. " +
                     "If this minimum has not been reached, bulk import jobs will refuse to start.")
             .defaultValue("64")
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
 
     UserDefinedInstanceProperty DEFAULT_INGEST_BATCHER_MIN_JOB_SIZE = Index.propertyBuilder("sleeper.default.ingest.batcher.job.min.size")
             .description("Specifies the minimum total file size required for an ingest job to be batched and sent. " +
@@ -192,7 +192,7 @@ public interface DefaultProperty {
                     "minimum number of files is also met.")
             .defaultValue("1G")
             .validationPredicate(SleeperPropertyValueUtils::isValidNumberOfBytes)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_BATCHER_MAX_JOB_SIZE = Index.propertyBuilder("sleeper.default.ingest.batcher.job.max.size")
             .description("Specifies the maximum total file size for a job in the ingest batcher. " +
                     "If more data is waiting than this, it will be split into multiple jobs. " +
@@ -201,14 +201,14 @@ public interface DefaultProperty {
                     "overflow the size of a job but aren't enough to create a job on their own.")
             .defaultValue("5G")
             .validationPredicate(SleeperPropertyValueUtils::isValidNumberOfBytes)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_BATCHER_MIN_JOB_FILES = Index.propertyBuilder("sleeper.default.ingest.batcher.job.min.files")
             .description("Specifies the minimum number of files for a job in the ingest batcher. " +
                     "An ingest job will be created if the batcher runs while this many files are waiting, and the " +
                     "minimum size of files is also met.")
             .defaultValue("1")
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_BATCHER_MAX_JOB_FILES = Index.propertyBuilder("sleeper.default.ingest.batcher.job.max.files")
             .description("Specifies the maximum number of files for a job in the ingest batcher. " +
                     "If more files are waiting than this, they will be split into multiple jobs. " +
@@ -216,7 +216,7 @@ public interface DefaultProperty {
                     "overflow the size of a job but aren't enough to create a job on their own.")
             .defaultValue("100")
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_BATCHER_MAX_FILE_AGE_SECONDS = Index.propertyBuilder("sleeper.default.ingest.batcher.file.max.age.seconds")
             .description("Specifies the maximum time in seconds that a file can be held in the batcher before it " +
                     "will be included in an ingest job. When any file has been waiting for longer than this, jobs " +
@@ -224,13 +224,13 @@ public interface DefaultProperty {
                     "met.")
             .defaultValue("300")
             .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_BATCHER_INGEST_QUEUE = Index.propertyBuilder("sleeper.default.ingest.batcher.ingest.queue")
             .description("Specifies the target ingest queue where batched jobs are sent.\n" +
                     "Valid values are: " + describeEnumValuesInLowerCase(IngestQueue.class))
             .defaultValue(IngestQueue.BULK_IMPORT_EMR_SERVERLESS.name().toLowerCase(Locale.ROOT))
             .validationPredicate(IngestQueue::isValid)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_BATCHER_TRACKING_TTL_MINUTES = Index.propertyBuilder("sleeper.default.ingest.batcher.file.tracking.ttl.minutes")
             .description("The time in minutes that the tracking information is retained for a file before the " +
                     "records of its ingest are deleted (eg. which ingest job it was assigned to, the time this " +
@@ -240,19 +240,19 @@ public interface DefaultProperty {
                     "Defaults to 1 week.")
             .defaultValue("" + 60 * 24 * 7)
             .validationPredicate(SleeperPropertyValueUtils::isNonNegativeInteger)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_FILE_WRITING_STRATEGY = Index.propertyBuilder("sleeper.default.ingest.file.writing.strategy")
             .description("Specifies the strategy that ingest uses to create files and references in partitions.\n" +
                     "Valid values are: " + describeEnumValuesInLowerCase(IngestFileWritingStrategy.class))
             .defaultValue(IngestFileWritingStrategy.ONE_REFERENCE_PER_LEAF.name().toLowerCase(Locale.ROOT))
             .validationPredicate(IngestFileWritingStrategy::isValid)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_RECORD_BATCH_TYPE = Index.propertyBuilder("sleeper.default.ingest.record.batch.type")
             .description("The way in which records are held in memory before they are written to a local store.\n" +
                     "Valid values are 'arraylist' and 'arrow'.\n" +
                     "The arraylist method is simpler, but it is slower and requires careful tuning of the number of records in each batch.")
             .defaultValue("arrow")
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_PARTITION_FILE_WRITER_TYPE = Index.propertyBuilder("sleeper.default.ingest.partition.file.writer.type")
             .description("The way in which partition files are written to the main Sleeper store.\n" +
                     "Valid values are 'direct' (which writes using the s3a Hadoop file system) and 'async' (which writes locally and then " +
@@ -260,7 +260,7 @@ public interface DefaultProperty {
                     "The direct method is simpler but the async method should provide better performance when the number of partitions " +
                     "is large.")
             .defaultValue("async")
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_ASYNC_COMMIT_BEHAVIOUR = Index.propertyBuilder("sleeper.default.statestore.commit.async.behaviour")
             .description("This is the default for whether state store updates will be applied asynchronously via the " +
                     "state store committer.\n" +
@@ -278,7 +278,7 @@ public interface DefaultProperty {
                     "overridden in table properties.")
             .defaultValue(DefaultAsyncCommitBehaviour.PER_IMPLEMENTATION.toString())
             .validationPredicate(DefaultAsyncCommitBehaviour::isValid)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_COMPACTION_JOB_ID_ASSIGNMENT_COMMIT_ASYNC = Index.propertyBuilder("sleeper.default.compaction.job.id.assignment.commit.async")
             .description("This is the default for whether created compaction jobs will be assigned to their input " +
                     "files asynchronously via the state store committer, if asynchronous commit is enabled. " +
@@ -286,40 +286,40 @@ public interface DefaultProperty {
                     "store.")
             .defaultValue("true")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_COMPACTION_JOB_COMMIT_ASYNC = Index.propertyBuilder("sleeper.default.compaction.job.commit.async")
             .description("This is the default for whether compaction tasks will commit finished jobs asynchronously " +
                     "via the state store committer, if asynchronous commit is enabled. Otherwise, compaction tasks " +
                     "will commit finished jobs directly to the state store.")
             .defaultValue("true")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_INGEST_FILES_COMMIT_ASYNC = Index.propertyBuilder("sleeper.default.ingest.job.files.commit.async")
             .description("This is the default for whether ingest tasks will add files asynchronously via the state " +
                     "store committer, if asynchronous commit is enabled. Otherwise, ingest tasks will add files " +
                     "directly to the state store.")
             .defaultValue("true")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_BULK_IMPORT_FILES_COMMIT_ASYNC = Index.propertyBuilder("sleeper.default.bulk.import.job.files.commit.async")
             .description("This is the default for whether bulk import will add files asynchronously via the state " +
                     "store committer, if asynchronous commit is enabled. Otherwise, bulk import will add files " +
                     "directly to the state store.")
             .defaultValue("true")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_PARTITION_SPLIT_ASYNC_COMMIT = Index.propertyBuilder("sleeper.default.partition.splitting.commit.async")
             .description("This is the default for whether partition splits will be applied asynchronously via the " +
                     "state store committer, if asynchronous commit is enabled. Otherwise, the partition splitter " +
                     "will apply splits directly to the state store.")
             .defaultValue("true")
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_GARBAGE_COLLECTOR_ASYNC_COMMIT = Index.propertyBuilder("sleeper.default.gc.commit.async")
             .description("This is the default for whether the garbage collector will record deleted files " +
                     "asynchronously via the state store committer, if asynchronous commit is enabled. Otherwise, the " +
                     "garbage collector will record this directly to the state store.")
             .defaultValue("true")
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_STATESTORE_COMMITTER_UPDATE_ON_EVERY_COMMIT = Index.propertyBuilder("sleeper.default.statestore.committer.update.every.commit")
             .description("When using the transaction log state store, this sets whether to update from the " +
                     "transaction log before adding a transaction in the asynchronous state store committer.\n" +
@@ -329,39 +329,14 @@ public interface DefaultProperty {
                     "be true to avoid conflicts and retries.")
             .defaultValue("false")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
     UserDefinedInstanceProperty DEFAULT_STATESTORE_COMMITTER_UPDATE_ON_EVERY_BATCH = Index.propertyBuilder("sleeper.default.statestore.committer.update.every.batch")
             .description("When using the transaction log state store, this sets whether to update from the " +
                     "transaction log before adding a batch of transactions in the asynchronous state store " +
                     "committer.")
             .defaultValue("true")
             .validationPredicate(SleeperPropertyValueUtils::isTrueOrFalse)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
-    UserDefinedInstanceProperty DEFAULT_LAMBDA_CONCURRENCY_RESERVED = Index.propertyBuilder("sleeper.default.lambda.concurrency.reserved")
-            .description("Default value for the reserved concurrency for each lambda within the Sleeper instance. " +
-                    "By default no concurrency is reserved for the lambdas. Each lambda also has its own property " +
-                    "that overrides the value found here.\n" +
-                    "See reserved concurrency overview at: https://docs.aws.amazon.com/lambda/latest/dg/configuration-concurrency.html")
-            .validationPredicate(SleeperPropertyValueUtils::isPositiveIntegerOrNull)
-            .defaultValue(null)
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
-    UserDefinedInstanceProperty DEFAULT_LAMBDA_CONCURRENCY_MAXIMUM = Index.propertyBuilder("sleeper.default.lambda.concurrency.max")
-            .description("Default value for the maximum concurrency for each lambda within the Sleeper instance. " +
-                    "By default the maximum concurrency is set to 10, which is enough for 10 online tables. " +
-                    "If there are more online tables, this number may need to be increased. Each lambda also has its own property that " +
-                    "overrides the value found here.\n" +
-                    "See maximum concurrency overview at: https://aws.amazon.com/blogs/compute/introducing-maximum-concurrency-of-aws-lambda-functions-when-using-amazon-sqs-as-an-event-source/")
-            .validationPredicate(SleeperPropertyValueUtils::isPositiveIntegerOrNull)
-            .defaultValue("10")
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
-    UserDefinedInstanceProperty DEFAULT_TABLE_STATE_LAMBDA_MEMORY = Index.propertyBuilder("sleeper.default.table.state.lambda.memory.mb")
-            .description("Default value for amount of memory in MB for each lambda that holds the state of Sleeper " +
-                    "tables in memory. These use a state store provider which caches a number of tables at " +
-                    "once, set in `sleeper.statestore.statestore.provider.cache.size`. Not all lambdas are covered " +
-                    "by this, e.g. see `sleeper.batch.table.lambdas.memory.mb`.")
-            .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .defaultValue("4096")
-            .propertyGroup(InstancePropertyGroup.DEFAULT).build();
+            .propertyGroup(InstancePropertyGroup.TABLE_PROPERTY_DEFAULT).build();
 
     static List<UserDefinedInstanceProperty> getAll() {
         return Index.INSTANCE.getAll();
