+++ java/core/src/test/java/sleeper/core/properties/instance/InstancePropertiesTest.java
@@ -77,7 +77,6 @@ import static sleeper.core.properties.instance.CompactionProperty.COMPACTION_TAS
 import static sleeper.core.properties.instance.CompactionProperty.COMPACTION_TASK_CREATION_PERIOD_IN_MINUTES;
 import static sleeper.core.properties.instance.CompactionProperty.ECR_COMPACTION_REPO;
 import static sleeper.core.properties.instance.CompactionProperty.MAXIMUM_CONCURRENT_COMPACTION_TASKS;
-import static sleeper.core.properties.instance.DefaultProperty.DEFAULT_LAMBDA_CONCURRENCY_RESERVED;
 import static sleeper.core.properties.instance.GarbageCollectionProperty.GARBAGE_COLLECTOR_LAMBDA_CONCURRENCY_RESERVED;
 import static sleeper.core.properties.instance.GarbageCollectionProperty.GARBAGE_COLLECTOR_LAMBDA_MEMORY_IN_MB;
 import static sleeper.core.properties.instance.GarbageCollectionProperty.GARBAGE_COLLECTOR_LAMBDA_TIMEOUT_IN_MINUTES;
@@ -234,7 +233,7 @@ class InstancePropertiesTest {
     void shouldGetTheDefaultPropertyWhenPropertyHasNotBeenSet() {
         // Given
         InstanceProperties properties = new InstanceProperties();
-        properties.set(DEFAULT_LAMBDA_CONCURRENCY_RESERVED, "10");
+        properties.set(CommonProperty.DEFAULT_LAMBDA_CONCURRENCY_RESERVED, "10");
 
         // When / Then
         assertThat(properties.get(GARBAGE_COLLECTOR_LAMBDA_CONCURRENCY_RESERVED)).isEqualTo("10");
